Refactor `PoolSv2`
Split library and binary code.
